stt_app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
annotated-doc=0.0.3=pyhcf101f3_1
annotated-types=0.7.0=pyhd8ed1ab_1
anyio=4.11.0=pyhcf101f3_0
aom=3.9.1=he0c23c2_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.11.4=haa95532_0
cairo=1.18.0=h32b962e_3
certifi=2025.10.5=pyhd8ed1ab_0
charset-normalizer=3.4.4=pypi_0
click=8.3.0=pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_1
dav1d=1.2.1=hcfcfb64_0
dnspython=2.8.0=pyhcf101f3_0
email-validator=2.3.0=pyhd8ed1ab_0
email_validator=2.3.0=hd8ed1ab_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
expat=2.7.3=h9214b88_0
fastapi=0.121.0=h67827d6_0
fastapi-cli=0.0.13=pyhcf101f3_0
fastapi-core=0.121.0=pyhcf101f3_0
ffmpeg=6.1.2=gpl_h9673905_705
filelock=3.20.0=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.14.1=h57928b3_0
fsspec=2025.10.0=pypi_0
graphite2=1.3.14=hac47afa_2
h11=0.16.0=pyhd8ed1ab_0
h2=4.3.0=pyhcf101f3_0
harfbuzz=9.0.0=h2bedf89_1
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pyh29332c3_0
httptools=0.7.1=py310h29418f3_0
httpx=0.28.1=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=75.1=he0c23c2_0
idna=3.11=pyhd8ed1ab_0
jinja2=3.1.6=pyhd8ed1ab_0
libffi=3.4.4=hd77b12b_1
libfreetype=2.14.1=h57928b3_0
libfreetype6=2.14.1=hdbac1cb_0
libglib=2.84.0=h7025463_0
libiconv=1.18=hc1393d2_2
libintl=0.22.5=h5728263_3
libopus=1.5.2=h2466b09_0
libpng=1.6.50=h7351971_1
libxml2=2.13.9=h741aa76_0
libzlib=1.3.1=h02ab6af_0
llvmlite=0.45.1=pypi_0
markdown-it-py=4.0.0=pyhd8ed1ab_0
markupsafe=3.0.3=py310hdb0e946_0
mdurl=0.1.2=pyhd8ed1ab_1
more-itertools=10.8.0=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.4.2=pypi_0
numba=0.62.1=pypi_0
numpy=2.2.6=pypi_0
openai-whisper=20250625=pypi_0
openh264=2.4.1=h63175ca_0
openssl=3.5.4=h725018a_0
pcre2=10.44=h99c9b8b_2
pip=25.2=pyhc872135_1
pixman=0.46.4=h5112557_1
portaudio=19.7.0=he0c23c2_0
pyaudio=0.2.11=py310h2bbff1b_2
pydantic=2.12.3=pyh3cfb1c2_0
pydantic-core=2.41.4=py310h034784e_0
pygments=2.19.2=pyhd8ed1ab_0
python=3.10.19=h981015d_0
python-dotenv=1.2.1=pyhcf101f3_0
python-multipart=0.0.20=pyhff2d567_0
python_abi=3.10=2_cp310
pyyaml=6.0.3=py310hdb0e946_0
regex=2025.11.3=pypi_0
requests=2.32.5=pypi_0
rich=14.2.0=pyhcf101f3_0
rich-toolkit=0.15.1=pyhcf101f3_0
setuptools=80.9.0=py310haa95532_0
shellingham=1.5.4=pyhd8ed1ab_1
sniffio=1.3.1=pyhd8ed1ab_1
sqlite=3.50.2=hda9a48d_1
starlette=0.49.2=pyhfdc7a7d_0
svt-av1=2.2.1=he0c23c2_0
sympy=1.14.0=pypi_0
tiktoken=0.12.0=pypi_0
tk=8.6.15=hf199647_0
torch=2.9.0=pypi_0
torchaudio=2.9.0=pypi_0
tqdm=4.67.1=pypi_0
typer=0.20.0=pyhdb1f59b_0
typer-slim=0.20.0=pyhcf101f3_0
typer-slim-standard=0.20.0=h65a100f_0
typing-extensions=4.15.0=h396c80c_0
typing-inspection=0.4.2=pyhd8ed1ab_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.5.0=pypi_0
uvicorn=0.38.0=pyh5737063_0
uvicorn-standard=0.38.0=h5737063_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
watchfiles=1.1.1=py310hf13f778_0
websockets=15.0.1=py310h1637853_2
wheel=0.45.1=py310haa95532_0
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xz=5.6.4=h4754444_1
yaml=0.2.5=h6a83c73_3
zlib=1.3.1=h02ab6af_0
